chore(dir): Переместить компоненты по специфичным пакетам
