+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/feature/shopping_lists/presentation/ShoppingListItem.kt
@@ -1,4 +1,4 @@
-package io.dimasla4ee.shoppinglist.components
+package io.dimasla4ee.shoppinglist.feature.shopping_lists.presentation
 
 import org.jetbrains.compose.resources.DrawableResource
